pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pgp-tools"
version = "0.1.0"
description = "CLI toolkit for managing GnuPG keyrings, trust, encryption, and recipient inspection"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "Maksym Tymoshyk", email = "maksym.tymoshyk@gmail.com" }]
keywords = ["pgp", "gpg", "gnupg", "encryption", "keyring", "trust", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "python-gnupg==0.5.2",
]

[project.optional-dependencies]
gcp = [
    "google-cloud-secret-manager>=2.27.0",
    "protobuf>=6.30.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.6.0",
    "pre-commit>=3.7.0",
    "build>=1.2.0",
    "twine>=5.0.0",
]

[project.scripts]
pgp-3des-cfb = "pgp_tools.pgp_3des_cfb:main"
pgp-list-recipients = "pgp_tools.list_recipients:main"
pgp-manage-trust = "pgp_tools.manage_trust:main"
pgp-fix-key-usability = "pgp_tools.fix_key_usability:main"
pgp-import-pubring = "pgp_tools.import_pubring:main"
pgp-import-secring = "pgp_tools.import_secring:main"
pgp-import-pubring-from-gcp = "pgp_tools.import_pubring_from_gcp:main"
pgp-import-secring-from-gcp = "pgp_tools.import_secring_from_gcp:main"
pgp-cleanup-secring = "pgp_tools.cleanup_secring:main"

[project.urls]
Homepage = "https://github.com/maksymtymoshyk/pgp-tools"
Issues = "https://github.com/maksymtymoshyk/pgp-tools/issues"
Changelog = "https://github.com/maksymtymoshyk/pgp-tools/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/pgp_tools"]

[tool.hatch.build.targets.sdist]
include = [
    "src/pgp_tools",
    "tests",
    "examples",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
    "pyproject.toml",
]

[tool.pytest.ini_options]
addopts = "-ra --strict-markers"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/pgp_tools"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.ruff]
line-length = 100
target-version = "py39"
src = ["src", "tests"]
extend-exclude = [".venv", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "PL", "RUF"]
ignore = [
    "PLR0913",
    "PLR2004",
    "PLR0911",
    "PLR0912",
    "PLR0915",
    "PLC0415",
    "PLW0108",
    "RUF059",
    "SIM105",
    "SIM108",
    "B008",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004", "S101", "F401", "F841", "RUF012"]
